+++ avr/cores/megacommand/GUI_private.cpp
@@ -154,7 +154,7 @@ ButtonsClass::ButtonsClass() {
 }
 
 void ButtonsClass::clear() {
-  for (int i = 0; i < GUI_NUM_BUTTONS; i++) {
+  for (uint8_t i = 0; i < GUI_NUM_BUTTONS; i++) {
     CLEAR_B_DOUBLE_CLICK(i);
     CLEAR_B_CLICK(i);
     CLEAR_B_LONG_CLICK(i);
